feat: Enhance logging

+++ src/mkmapdiary/util/os.py
@@ -1,10 +1,14 @@
 import pathlib
 import shutil
+from typing import Optional, List
 
 
-def clean_dir(build_dir: pathlib.Path):
+def clean_dir(build_dir: pathlib.Path, keep_files: Optional[List[str]] = None):
+    if keep_files is None:
+        keep_files = []
     for item in build_dir.iterdir():
         if item.is_dir():
             shutil.rmtree(item)
         else:
-            item.unlink()
+            if item.name not in keep_files:
+                item.unlink()
